+++ src/data/records/sam/reader.rs
@@ -5,7 +5,11 @@ use std::{
 };
 
 use crate::{
-    data::{cigar::Cigar, records::RecordReader, sam::SamData},
+    data::{
+        cigar::Cigar,
+        records::RecordReader,
+        sam::{SamData, SamTags},
+    },
     unwrap_or_return_some_err,
 };
 
@@ -56,11 +60,11 @@ impl std::fmt::Display for SamRow {
 /// An iterator for buffered reading of a SAM file. Guarantees quality scores
 /// are valid.
 #[derive(Debug)]
-pub struct SAMReader<R: std::io::Read> {
+pub struct SAMReader<R: std::io::Read, const TAGS: bool> {
     pub sam_reader: std::io::Lines<std::io::BufReader<R>>,
 }
 
-impl<R: std::io::Read> SAMReader<R> {
+impl<R: std::io::Read> SAMReader<R, true> {
     /// Creates an iterator over SAM data, wrapping the input in a buffered
     /// reader.
     ///
@@ -68,7 +72,12 @@ impl<R: std::io::Read> SAMReader<R> {
     /// initially. It also allows for empty input, in which case the resulting
     /// iterator is empty.
     ///
+    /// If the optional tags in the SAM records are not being used by the
+    /// downstream application, consider using [`new_ignore_tags`] for
+    /// efficiency.
+    ///
     /// [`from_readable`]: SAMReader::from_readable
+    /// [`new_ignore_tags`]: SAMReader::new_ignore_tags
     pub fn new(inner: R) -> Self {
         SAMReader {
             sam_reader: std::io::BufReader::new(inner).lines(),
@@ -78,11 +87,16 @@ impl<R: std::io::Read> SAMReader<R> {
     /// Creates an iterator over SAM data from a type implementing [`Read`],
     /// wrapping the input in a buffered reader.
     ///
+    /// If the optional tags in the SAM records are not being used by the
+    /// downstream application, consider using [`from_readable_ignore_tags`] for
+    /// efficiency.
+    ///
     /// ## Errors
     ///
     /// Will return `Err` if the input data is empty or an IO error occurs.
     ///
     /// [`Read`]: std::io::Read
+    /// [`from_readable_ignore_tags`]: SAMReader::from_readable_ignore_tags
     pub fn from_readable(read: R) -> std::io::Result<Self> {
         let mut sam_reader = std::io::BufReader::new(read);
         if sam_reader.fill_buf()?.is_empty() {
@@ -95,96 +109,163 @@ impl<R: std::io::Read> SAMReader<R> {
     }
 }
 
-impl<R: std::io::Read> Iterator for SAMReader<R> {
+impl<R: std::io::Read> SAMReader<R, false> {
+    /// Creates an iterator over SAM data, wrapping the input in a buffered
+    /// reader.
+    ///
+    /// This version ignores any optional tags present in the SAM records in
+    /// order to reduce allocations and improve efficiency. Use [`new`] to
+    /// include the tags in the output.
+    ///
+    /// Unlike [`from_readable_ignore_tags`], this does not allocate or read any
+    /// data initially. It also allows for empty input, in which case the
+    /// resulting iterator is empty.
+    ///
+    /// [`from_readable_ignore_tags`]: SAMReader::from_readable_ignore_tags
+    /// [`new`]: SAMReader::new
+    pub fn new_ignore_tags(inner: R) -> Self {
+        SAMReader {
+            sam_reader: std::io::BufReader::new(inner).lines(),
+        }
+    }
+
+    /// Creates an iterator over SAM data from a type implementing [`Read`],
+    /// wrapping the input in a buffered reader.
+    ///
+    /// This version ignores any optional tags present in the SAM records in
+    /// order to reduce allocations and improve efficiency. Use
+    /// [`from_readable`] to include the tags in the output.
+    ///
+    /// ## Errors
+    ///
+    /// Will return `Err` if the input data is empty or an IO error occurs.
+    ///
+    /// [`Read`]: std::io::Read
+    /// [`from_readable`]: SAMReader::from_readable
+    pub fn from_readable_ignore_tags(read: R) -> std::io::Result<Self> {
+        let mut sam_reader = std::io::BufReader::new(read);
+        if sam_reader.fill_buf()?.is_empty() {
+            return Err(IOError::new(ErrorKind::InvalidData, "No SAM data was found!"));
+        }
+
+        Ok(SAMReader {
+            sam_reader: sam_reader.lines(),
+        })
+    }
+}
+
+impl<R: std::io::Read, const TAGS: bool> Iterator for SAMReader<R, TAGS> {
     type Item = std::io::Result<SamRow>;
 
     fn next(&mut self) -> Option<Self::Item> {
-        let line = self.sam_reader.next()?;
+        let line = unwrap_or_return_some_err!(self.sam_reader.next()?);
 
-        match line {
-            Ok(s) if s.starts_with('@') => Some(Ok(SamRow::Header(s))),
-            Ok(s) => {
-                let r: Vec<&str> = s.split('\t').collect();
-                // TODO: account for optional fields
-                if r.len() < 11 {
-                    return Some(Err(IOError::new(
-                        ErrorKind::InvalidData,
-                        "SAM file did not have at least 11 fields!",
-                    )));
-                }
+        if line.starts_with('@') {
+            Some(Ok(SamRow::Header(line)))
+        } else {
+            let parts = line.split('\t').collect::<Vec<_>>();
 
-                let read_bit_flags = match r[1].parse::<u16>() {
+            if parts.len() < 11 {
+                return Some(Err(IOError::new(
+                    ErrorKind::InvalidData,
+                    "SAM file did not have at least 11 fields!",
+                )));
+            }
+
+            let qname = parts[0].to_string();
+
+            let flag = {
+                match parts[1].parse::<u16>() {
                     Ok(f) => f,
                     Err(e) => {
                         return Some(Err(IOError::new(
                             ErrorKind::InvalidData,
-                            format!("Error: {e}, invalid SAM bit flag '{}'", &r[1]),
+                            format!("Error: {e}, invalid SAM bit flag '{}'", parts[1]),
                         )));
                     }
-                };
+                }
+            };
 
-                // Per the SAM spec, the 1-based reference position
-                let aligned_reference_position = match r[3].parse::<usize>() {
+            let rname = parts[2].to_string();
+
+            let pos = {
+                match parts[3].parse::<usize>() {
                     Ok(p) => p,
                     Err(e) => {
                         return Some(Err(IOError::new(
                             ErrorKind::InvalidData,
-                            format!("Error: {e}, invalid SAM reference position '{}'", &r[3]),
+                            format!("Error: {e}, invalid SAM reference position '{}'", parts[3]),
                         )));
                     }
-                };
+                }
+            };
 
-                let read_mapping_quality = match r[4].parse::<u8>() {
+            let mapq = {
+                match parts[4].parse::<u8>() {
                     Ok(q) => q,
                     Err(e) => {
                         return Some(Err(IOError::new(
                             ErrorKind::InvalidData,
-                            format!("Error: {e}, invalid SAM mapq '{}'", &r[4]),
+                            format!("Error: {e}, invalid SAM mapq '{}'", parts[4]),
                         )));
                     }
-                };
+                }
+            };
 
-                let quality_scores = unwrap_or_return_some_err!(r[10].as_bytes().try_into());
+            let cigar = {
+                let cigar_str = parts[5].trim_ascii().as_bytes();
+                if cigar_str == b"*" {
+                    Cigar::new()
+                } else {
+                    Cigar::from_slice_unchecked(cigar_str)
+                }
+            };
 
-                let cigar = {
-                    let out = r[5].trim_ascii().as_bytes();
-                    if out == b"*" {
-                        Cigar::new()
-                    } else {
-                        Cigar::from_slice_unchecked(out)
-                    }
-                };
+            // TODO: Where should casing be handled?
+            let seq = parts[9].as_bytes().to_ascii_uppercase().into();
 
-                let row = SamData {
-                    qname: r[0].to_owned(),
-                    flag: read_bit_flags,
-                    rname: r[2].to_owned(),
-                    pos: aligned_reference_position,
-                    mapq: read_mapping_quality,
-                    cigar,
-                    rnext: '*',
-                    pnext: 0,
-                    tlen: 0,
-                    // TODO: Where should casing be handled?
-                    seq: r[9].as_bytes().to_ascii_uppercase().into(),
-                    qual: quality_scores,
-                };
-                Some(Ok(SamRow::Data(row)))
-            }
-            Err(e) => Some(Err(e)),
+            let qual = unwrap_or_return_some_err!(parts[10].as_bytes().try_into());
+
+            let tags = if TAGS {
+                parts[11..].iter().map(ToString::to_string).collect::<SamTags>()
+            } else {
+                SamTags::new()
+            };
+
+            let row = SamData {
+                qname,
+                flag,
+                rname,
+                pos,
+                mapq,
+                cigar,
+                rnext: '*',
+                pnext: 0,
+                tlen: 0,
+                seq,
+                qual,
+                tags,
+            };
+            Some(Ok(SamRow::Data(row)))
         }
     }
     // Push down predicates?
 }
 
-impl SAMReader<std::fs::File> {
+impl SAMReader<File, true> {
     /// Reads a SAM text file into an iterator backed by a buffered reader.
+    ///
+    /// If the optional tags in the SAM records are not being used by the
+    /// downstream application, consider using [`from_filename_ignore_tags`] for
+    /// efficiency.
     ///
     /// ## Errors
     ///
     /// Will return `Err` if file or permissions do not exist, or if the file is
     /// empty. The file path is included in the error message.
-    pub fn from_filename<P>(filename: P) -> Result<SAMReader<File>, std::io::Error>
+    ///
+    /// [`from_filename_ignore_tags`]: SAMReader::from_filename_ignore_tags
+    pub fn from_filename<P>(filename: P) -> Result<Self, std::io::Error>
     where
         P: AsRef<Path>, {
         let file = Self::open_nonempty_file(filename)?;
@@ -192,6 +273,27 @@ impl SAMReader<std::fs::File> {
     }
 }
 
-impl<R: std::io::Read> RecordReader for SAMReader<R> {
+impl SAMReader<File, false> {
+    /// Reads a SAM text file into an iterator backed by a buffered reader.
+    ///
+    /// This version ignores any optional tags present in the SAM records in
+    /// order to reduce allocations and improve efficiency. Use [`from_filename`] to
+    /// include the tags in the output.
+    ///
+    /// ## Errors
+    ///
+    /// Will return `Err` if file or permissions do not exist, or if the file is
+    /// empty. The file path is included in the error message.
+    ///
+    /// [`from_filename`]: SAMReader::from_filename
+    pub fn from_filename_ignore_tags<P>(filename: P) -> Result<Self, std::io::Error>
+    where
+        P: AsRef<Path>, {
+        let file = Self::open_nonempty_file(filename)?;
+        Ok(SAMReader::new_ignore_tags(file))
+    }
+}
+
+impl<R: std::io::Read, const TAGS: bool> RecordReader for SAMReader<R, TAGS> {
     const RECORD_NAME: &str = "SAM";
 }