+++ src/data/records/sam/std_traits.rs
@@ -15,12 +15,17 @@ impl std::fmt::Display for SamData {
             tlen,
             seq,
             qual,
+            tags,
         } = self;
 
         write!(
             f,
             "{qname}\t{flag}\t{rname}\t{pos}\t{mapq}\t{cigar}\t{rnext}\t{pnext}\t{tlen}\t{seq}\t{qual}"
-        )
+        )?;
+        for tag in &tags.0 {
+            write!(f, "\t{tag}")?;
+        }
+        Ok(())
     }
 }